refactor!: rename dscrbo to reveal
"dscrbo" read as a typo of its sibling "describo", which made every discussion
of the two need a disambiguating clause. The libraries are independent and
should be recognisable apart at a glance.

Directory, Scala package and artifact all move together:

  libs/commons/dscrbo   -> libs/commons/reveal
  com.worxbend.dscrbo   -> com.worxbend.reveal
  dscrbo_3              -> reveal_3

Publishing organization is unchanged. The mixin's members follow the package,
so AutoToString now declares revealDescribe and revealConfiguration; they stay
prefixed for the reason they always were, that a one-letter name would collide
with a user's field.

Two things a textual rename does not catch, both found by the test suite rather
than assumed:

* Five expectations pin the *compressed* package prefix, which is derived from
  the package name rather than spelled out in it. `com.worxbend.dscrbo` under
  shortPackagePrefix rendered `c.w.d`, and now renders `c.w.r`. describo's own
  `c.w.d` expectations are deliberately untouched.
* The README's test badge still pointed at the conformance-kit section that was
  removed in the previous commit, and carried a stale count.

describo's README no longer frames the enum-naming difference as "one of the two
places they disagree", which was left over from the parity contract; the
libraries are independent, so neither is obliged to follow the other.

No behaviour change. 403 tests (describo 186, reveal 217) all pass, the module
compiles warning-free under -Wunused:all, and no reference to the old name
remains anywhere in the repository.
