+++ libs/commons/describo/README.md
@@ -7,9 +7,9 @@ hand-written `toString`, no reflection at render time, and no way for a `@Redact
 
 - Scala 3 only (3.8.4).
 - One dependency: `com.softwaremill.magnolia1_3::magnolia`.
-- Sibling module [`dscrbo`](../dscrbo/README.md) solves the same problem with an inline macro and
+- Sibling module [`reveal`](../reveal/README.md) solves the same problem with an inline macro and
   *zero* dependencies. The two are **independent and are not required to render identically**. Pick
-  describo when you want typeclass composition and instances for third-party types; pick dscrbo when
+  describo when you want typeclass composition and instances for third-party types; pick reveal when
   you cannot take a dependency.
 
 ## Coordinates
@@ -24,7 +24,7 @@ That is intentional, not an oversight: `io.worxbend` is this repository's establ
 `com.worxbend.<product>.<concept>` is the mandated package prefix for new code. Please do not
 "fix" one to match the other.
 
-`dscrbo` deliberately lives in a **different leaf package**, `com.worxbend.dscrbo`. Both artifacts
+`reveal` deliberately lives in a **different leaf package**, `com.worxbend.reveal`. Both artifacts
 define `Configuration`, `annotations.Redacted` and `annotations.Excluded`, so distinct leaf
 packages are what makes the two co-installable on one classpath.
 
@@ -106,7 +106,7 @@ Consequences, all of them deliberate:
   `@Excluded` is the intended spelling for new code.
 - Several `@Redacted` annotations on one field are not an error: **the one written first in source
   order wins**. Magnolia surfaces `param.annotations` in reverse source order, so `FieldRule.of`
-  reverses before searching; without that, this module and `dscrbo` — whose macro sorts by
+  reverses before searching; without that, this module and `reveal` — whose macro sorts by
   `pos.start` and takes the head — would disagree on the same input. Deterministic, and identical in
   both modules.
 
@@ -290,21 +290,21 @@ value-class parameter cannot omit anything — there would be nothing left to pr
   `assertDoesNotCompile`, so the day it changes the test fails rather than the behaviour drifting.
 - **Enum cases under `fullyQualifiedClassName`** carry Magnolia's `TypeInfo`, which reports the
   enclosing *package* rather than the enclosing enum, so `Colour.Red` prints as
-  `com.example.Red`. Simple names are unaffected. **This is one of the two places `describo` and
-  `dscrbo` disagree**: the sibling macro reads the case symbol directly and prints
-  `com.example.Colour.Red`, which is the more useful spelling. Fixing it here would mean threading the
+  `com.example.Red`. Simple names are unaffected. `reveal` spells this differently — its macro reads
+  the case symbol directly and prints `com.example.Colour.Red`, which is the more useful of the two —
+  but the libraries are independent, so neither is obliged to follow the other. Fixing it here would mean threading the
   parent's name through `split` into every child instance — a change to the `Printable` interface for a
   spelling that only appears under a non-default flag. Both behaviours are pinned by
   `NamingAndGenericsSuite`.
 - **Set and Map iteration order** is the insertion order only up to four elements; beyond that
   Scala switches to a hashed representation. Sort before rendering if you need stable output.
 
-## How this module differs from `dscrbo`
+## How this module differs from `reveal`
 
 The two are independent implementations and are **not** held to identical output. The differences worth
 knowing about:
 
-| | `describo` | `dscrbo` |
+| | `describo` | `reveal` |
 | :-- | :-- | :-- |
 | Nested case class, no instance | auto-derived by magnolia | plain `toString`; nested types are never unrolled |
 | Enum case, qualified name | `com.example.Red` | `com.example.Colour.Red` |
@@ -339,8 +339,8 @@ made together, on purpose, before the first release:
   exception to this repository's `noDefaultArgs` guidance. It is an options DTO consumed with named
   arguments — the one shape where booleans are self-documenting at the call site — and a sixteen-field
   record without defaults would be unusable. It is also duplicated field-for-field in
-  `com.worxbend.dscrbo`, which cannot depend on this module; any ADT would have to be duplicated too.
-  **Any change to `Configuration` must be mirrored in dscrbo and in both READMEs.**
+  `com.worxbend.reveal`, which cannot depend on this module; any ADT would have to be duplicated too.
+  **Any change to `Configuration` must be mirrored in reveal and in both READMEs.**
 - The genuine decisions are modelled as ADTs, privately: `FieldRule` (`Omit` / `Redact` / `Render`)
   and `Layout` (`SingleLine` / `Multiline`). The `<= 0` sentinel of
   `multilineIfFieldsAreGreaterOrEqual` is laundered into a `Layout` at one place rather than being
@@ -348,5 +348,5 @@ made together, on purpose, before the first release:
 - Annotation scanning happens once per typeclass instance, at derivation time, and is materialised
   as a `Vector` — not a lazy `View` whose filter re-runs on every `size` and every `map`.
 - The build enables `-Wunused:all -deprecation -feature` and mixes in Mill's `ScalafmtModule`, the
-  same as `dscrbo`, so dead code and hand-formatting are caught by `./mill libs.commons.describo.compile`
+  same as `reveal`, so dead code and hand-formatting are caught by `./mill libs.commons.describo.compile`
   and `./mill libs.commons.describo.checkFormat` rather than by review.
