+++ libs/commons/reveal/src/test/scala/com/worxbend/reveal/AutoToStringSuite.scala
@@ -1,6 +1,6 @@
-package com.worxbend.dscrbo
+package com.worxbend.reveal
 
-import com.worxbend.dscrbo.annotations.Redacted
+import com.worxbend.reveal.annotations.Redacted
 
 import org.scalatest.funsuite.AnyFunSuite
