+++ libs/commons/reveal/src/test/scala/com/worxbend/reveal/NestingDepthSuite.scala
@@ -1,4 +1,4 @@
-package com.worxbend.dscrbo
+package com.worxbend.reveal
 
 import scala.compiletime.testing.typeCheckErrors
