+++ libs/commons/reveal/src/test/scala/com/worxbend/reveal/SealedFamilySuite.scala
@@ -1,6 +1,6 @@
-package com.worxbend.dscrbo
+package com.worxbend.reveal
 
-import com.worxbend.dscrbo.annotations.Redacted
+import com.worxbend.reveal.annotations.Redacted
 
 import scala.util.Failure
 import scala.util.Success
